Look up the store for an address book dynamically
Previously, there was only one store, so the code assumed that that store was
used. Now we want to use a second store, the public store, so it becomes
necessary to lookup the store when given an address book ID. We do this by
saving the store when we look up the address books.