If no PR_CARDDAV_URI available, generate one from the EntryID
Previously the code generated a random GUID and wrote it back to the contact,
but that does not work if it's a public contact made by someone else, since we
don't have write privileges for it. So instead, generate an URI from the
EntryID, which is unique and hopefully sufficiently stable.