class.zarafastore.php: add Zarafa_Store class
Previously the code knew about a 'private store' and a 'public store', storing
data for both in global variables inside the Bridge. Now create a Zarafa_Store
class that encapsulates all knowledge about a store, and have the Bridge create
instances of this class for all private and public stores.

This provides encapsulation, modularization and extensibility.