Remove PHP closing tag from all PHP-only files
There is no reason to include the closing tag for a file containing only PHP,
since it is implied by the EOF. But including it can lead to printing
extraneous whitespace if there are trailing characters after the closing tag.

This also fixes the issue of missing/present newlines once and for all.