ZarafaBridge: logging: use proper facility
Trace is lowest, then debug, info, warn, error and fatal.

Entering a function warrants a trace statement. Performing an action is a debug
or an info statement. Errors in the "not found" category are warnings. Errors
in the "unexpected thing happened" category are errors. Errors in the "I can't
continue" category are fatal.