Fix Issue 25: 'continue' only breaks switch, not foreach
More information: http://code.google.com/p/sabre-zarafa/issues/detail?id=25

Issue 25 at Google Code rightly points out that 'continue' in a switch
statement is equivalent to 'break'. It doesn't continue to the next foreach()
iteration; for that we need 'continue 2'.

The issue with 'debug' mentioned in the ticket was fixed previously.