SabreDAV: remove bundled copy
SabreDAV 1.8 has a different directory layout from 1.6, which makes it
infeasible to include parts of the package, like we did with the 1.6 branch.
For instance the /vendor directory contains autoloader.php, which we need to
autoload the SabreDAV classes. A good reason to remove SabreDAV entirely and
require the end user to download and install it separately.