ZarafaFolder: always calculate CTag from all child cards
Previously, the CTag was calculated from the folder's last modification time if
available. It turns out that this time is not updated whenever a child is
edited in-place. To produce an accurate CTag, and thus to make clients update
properly, we need to look at the modification times on all the child cards.
This is unfortunate but seems inevitable.