server.php: hide errors from user, don't log username/pass
Change the error behaviour from throwing an uncaught ErrorException to logging
the error and quitting the script. Previously, the ErrorException would include
a stack trace, which in the case of a failed auth attempt, would include the
username and password. This ErrorException was not caught, and would end up in
the logs somewhere. Clearly not ideal.

Also change the 'display_errors' ini variable to not display errors by default;
they are written to the logs and that should be enough.