Zarafa_Folder: decrease log level in get_dav_card()
Not being able to find the requested contact in get_dav_card() is technically
an error, but log it at the low INFO log setting. Some clients, notably Mac
OSX' Contacts.app, test whether the name of a new card is available by
requesting that card and *expecting* the lookup to fail. So this "error" occurs
during normal operation. Don't pollute the logs unnecessarily.