VCardParser: don't use versions, encapsulate variables
The code was split into different classes for different Sabre-VObject versions,
but none of the versions differed much in practice. Write against Sabre-VObject
3.x, which is the current version bundled with SabreDAV 1.8.

Change the class interfaces to not pass the 'vCard' reader object around by
reference; make it a class member variable. This makes it clearer what the code
is doing.