README.md: change default OSX port, stress importance of SSL
As noted on the Zarafa forums, the default port for OSX Addressbook is not 8008
but 8843. Update the documentation to mention this fact. Also stress the
importance of using SSL. Though it was noted further on in the README, it's
really something that should be mentioned in the webserver config section.