Don't interpret the REV: field as a Unix timestamp
Evolution caused Sabre-Zarafa to blow up by not sending Unix timestamps in the
REV: field, but MD5 hashes. Remove read support for the REV: property entirely,
because we don't know how to interpret it, and we can't save it to MAPI (the
last-updated timestamp is set by the database layer, outside our control).

Keep write support, since it conforms to RFC 2426, see:
  http://tools.ietf.org/html/rfc2426#section-3.6.4