VCardProducer/VCardParser: upgrade to Sabre-VObject 3.x
Sabre-VObject 3.x makes a couple of non-backwards-compatible changes; update
the code to work with the new version. Sabre-Zarafa will now no longer work
with Sabre-VObject 2.x, which is the version currently bundled by SabreDAV.

The big reason to change to Sabre-VObject 3.x as soon as possible is that it
fixes the escaping bugs that are rampant in version 2.x. In particular,
newlines in notes were breaking the sync on OS X for a lot of people.