VCardProducer: SORT-AS property should be SORT-STRING
The SORT-AS property is a vCard 4.0 invention; in vCard 3.0, the property is
called SORT-STRING. Add code to generate the SORT-AS parameter to the N
property for vCard-4.0 cards as specified in the RFC; the catch is that the
underlying Sabre-VObject 3.0.0-alpha4 library cannot properly escape the value
yet (embed it in double quotes). Hope this doesn't break anything - the default
is not to create version 4.0 cards.