Remove default error handler, see issue #12
The default error handler was invoked when mapi_logon_zarafa() threw a login
error, and would quit the program. The desired behaviour is not to quit the
program, but to send a '401 Auth Required' response back to the client. Because
this handler was interfering, and because it does not have a good reason to
exist (if anything, we should set a handler for unhandled *exceptions*), we're
removing it.

In the future, we could consider enabling MAPI exceptions, and inserting a
default handler for uncaught MAPI exceptions.