server.php is responsible for loading MAPI libs and config
Since server.php is always the entry point for our application, make it
responsible for loading the MAPI libs, the config file, the version defines,
and so on. Also define our custom MAPI properties here, so that they are
available to all classes further on in the request.

This obsoletes 'common.inc.php'.